requirements.txt
# Web框架和服务器
fastapi>=0.115.0
uvicorn>=0.22.0
python-multipart>=0.0.5
jinja2>=3.1.2
starlette>=0.35.0
aiofiles>=23.2.1

# 图像和视频处理
opencv-python>=4.7.0
Pillow>=10.0.0

# HTTP客户端
httpx>=0.24.0

# AI模型集成
ollama>=0.2.0
google-generativeai==0.3.1
dashscope>=1.10.0
openai>=1.0.0

# 工具和辅助库
python-jose[cryptography]>=3.3.0
tenacity>=8.2.3
python-dotenv>=0.19.0
pydantic>=2.0.0

# 异步支持
aiohttp>=3.8.0
asyncio>=3.4.3

# 开发工具
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

# 其他依赖
tqdm>=4.66.0